Add accountInfo back to DonationBanner and merge old banner
