+++ app/web/features/donations/DonationBanner.tsx
@@ -1,5 +1,12 @@
 import VolunteerActivismIcon from "@mui/icons-material/VolunteerActivism";
-import { Alert, alpha, Button, styled } from "@mui/material";
+import {
+  Alert,
+  alpha,
+  Box,
+  Button,
+  LinearProgress,
+  styled,
+} from "@mui/material";
 import { useTranslation } from "i18n";
 import { GLOBAL } from "i18n/namespaces";
 import { useRouter } from "next/router";
@@ -9,19 +16,74 @@ import { donationsRoute } from "routes";
 import { theme } from "theme";
 
 import useAccountInfo from "../auth/useAccountInfo";
+import useDonationStats from "./useDonationStats";
 
 const TIME_BETWEEN_NAGS_MS = 24 * 60 * 60 * 1000; // 24 hours
 
-const Wrapper = styled("div")(({ theme }) => ({
+const OuterWrapper = styled("div")(({ theme }) => ({
   display: "flex",
-  justifyContent: "space-between",
   alignItems: "center",
   width: "100%",
   gap: theme.spacing(2),
   [theme.breakpoints.down("md")]: {
     flexDirection: "column",
-    alignItems: "center",
-    gap: theme.spacing(1),
+    alignItems: "stretch",
+    gap: theme.spacing(1.5),
+  },
+}));
+
+const ContentWrapper = styled("div")(({ theme }) => ({
+  display: "flex",
+  flexDirection: "column",
+  flex: 1,
+  gap: theme.spacing(0.75),
+}));
+
+const ThermometerRow = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  gap: theme.spacing(1.5),
+  width: "100%",
+}));
+
+const StyledLinearProgress = styled(LinearProgress)(({ theme }) => ({
+  height: 14,
+  borderRadius: 7,
+  flexGrow: 1,
+  backgroundColor: alpha(theme.palette.secondary.main, 0.15),
+  "& .MuiLinearProgress-bar": {
+    borderRadius: 7,
+    background: `linear-gradient(90deg, ${theme.palette.secondary.main} 0%, ${theme.palette.secondary.light} 100%)`,
+  },
+}));
+
+const ProgressLabel = styled("span")(({ theme }) => ({
+  fontSize: "0.875rem",
+  fontWeight: 600,
+  color: theme.palette.secondary.main,
+  whiteSpace: "nowrap",
+}));
+
+const Message = styled("span")(({ theme }) => ({
+  fontSize: "0.875rem",
+  color: theme.palette.text.secondary,
+  [theme.breakpoints.down("md")]: {
+    textAlign: "center",
+  },
+}));
+
+const StyledButton = styled(Button)(({ theme }) => ({
+  backgroundColor: theme.palette.secondary.main,
+  flexShrink: 0,
+  alignSelf: "center",
+  paddingLeft: theme.spacing(3),
+  paddingRight: theme.spacing(3),
+  "&:hover": {
+    backgroundColor: theme.palette.secondary.dark,
+  },
+  [theme.breakpoints.down("md")]: {
+    width: "100%",
+    alignSelf: "stretch",
   },
 }));
 
@@ -29,14 +91,23 @@ export function DonationBanner() {
   const { t } = useTranslation(GLOBAL);
   const router = useRouter();
 
-  // the epoch value of the last time this banner was dismissed
   const [lastDismissedEpoch, setLastDismissedEpoch] = usePersistedState<
     number | null
   >("donation_banner.dismissed", null);
   const [bannerVisible, setBannerVisible] = useState<boolean>(false);
 
+  const { data: apiDonationStats, isLoading } = useDonationStats();
   const { data: accountInfo, isLoading: isAccountInfoLoading } =
     useAccountInfo();
+
+  useEffect(() => {
+    if (!isLoading && apiDonationStats) {
+      setBannerVisible(
+        !lastDismissedEpoch ||
+          new Date().getTime() - lastDismissedEpoch > TIME_BETWEEN_NAGS_MS,
+      );
+    }
+  }, [isLoading, apiDonationStats, lastDismissedEpoch]);
 
   useEffect(() => {
     if (!isAccountInfoLoading && accountInfo?.shouldShowDonationBanner) {
@@ -59,7 +130,15 @@ export function DonationBanner() {
     setBannerVisible(false);
   };
 
-  if (!bannerVisible) return null;
+  if (!bannerVisible || !apiDonationStats) return null;
+
+  const progress = Math.min(
+    (apiDonationStats.totalDonatedYtd / apiDonationStats.goal) * 100,
+    100,
+  );
+
+  const formattedRaised = apiDonationStats.totalDonatedYtd.toLocaleString();
+  const formattedGoal = apiDonationStats.goal.toLocaleString();
 
   return (
     <Alert
@@ -67,31 +146,41 @@ export function DonationBanner() {
       onClose={handleClose}
       sx={{
         alignItems: "center",
-        ".MuiAlert-message": { width: "100%" },
+        ".MuiAlert-message": { width: "100%", py: 0.5 },
         backgroundColor: alpha(theme.palette.secondary.main, 0.08),
         color: theme.palette.text.primary,
         "& .MuiAlert-icon": {
           color: theme.palette.secondary.main,
+          alignSelf: "center",
+        },
+        [theme.breakpoints.down("md")]: {
+          alignItems: "flex-start",
+          "& .MuiAlert-icon": {
+            paddingTop: "10px",
+          },
         },
       }}
     >
-      <Wrapper>
-        <span>{t("donation_banner.message")}</span>
-        <Button
-          variant="contained"
-          size="small"
-          sx={{
-            backgroundColor: theme.palette.secondary.main,
-            flexShrink: 0,
-            "&:hover": {
-              backgroundColor: theme.palette.secondary.dark,
-            },
-          }}
-          onClick={handleDonateClick}
-        >
+      <OuterWrapper>
+        <ContentWrapper>
+          <ThermometerRow>
+            <Box sx={{ flexGrow: 1, minWidth: 0 }}>
+              <StyledLinearProgress
+                variant="determinate"
+                value={progress}
+                aria-label={t("donation_banner.progress_label")}
+              />
+            </Box>
+            <ProgressLabel>
+              ${formattedRaised} / ${formattedGoal}
+            </ProgressLabel>
+          </ThermometerRow>
+          <Message>{t("donation_banner.message")}</Message>
+        </ContentWrapper>
+        <StyledButton variant="contained" onClick={handleDonateClick}>
           {t("donation_banner.button")}
-        </Button>
-      </Wrapper>
+        </StyledButton>
+      </OuterWrapper>
     </Alert>
   );
 }
